requirements.txt
pyftpdlib==1.5.9
requests==2.31.0
python-dotenv==1.0.0
pyyaml==6.0.1 